knn with hyperparameter tuning
Improved the model doing hyperparameter tuning using GridSearchCV with 10-fold crossvalidation (tests exhaustively combinations of parameters in a given range). Now the best result of a 10-fold crossvalidation gives 0.15604737 misclassification.
Increased the features to include 'Difference in words lead and co-lead', 'Number words female', as well as the 2 composite features.
Also tried a couple of approaches for feature evaluation, but didn't seem to give relevant values.. so for the moment doing feature selection by hand..